Tiny fix

+++ blockchain/chaindata.go
@@ -251,6 +251,10 @@ func (c *ChainStoreExtend) renewCrCandidates() {
 		if err != nil {
 			return
 		}
+		cands := c.chain.GetCRCommittee().GetAllCandidates()
+		if len(cands) == 0 {
+			return
+		}
 		stmt1, err := db.Prepare("delete from chain_cr_candidate_info")
 		if err != nil {
 			return
@@ -265,7 +269,6 @@ func (c *ChainStoreExtend) renewCrCandidates() {
 		if err != nil {
 			return
 		}
-		cands := c.chain.GetCRCommittee().GetAllCandidates()
 		for i, can := range cands {
 			did, _ := can.Info().CID.ToAddress()
 			_, err = stmt.Exec(hex.EncodeToString(can.Info().Code), did, can.Info().NickName, can.Info().Url, can.Info().Location, can.State().String(), can.Votes().String(), i)
